@@ -8,26 +8,26 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3",
         "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^6.5",
-        "laravel/framework": "^7.25",
+        "guzzlehttp/guzzle": "^7.1.1",
+        "laravel/framework": "^8.9",
         "laravel/tinker": "^2.4",
-        "laravel/passport": "^9.3",
-        "laravel/ui": "^2.1",
-        "doctrine/dbal": "^2.10",
-        "laraveldaily/laravel-invoices": "^1.3",
-        "yajra/laravel-datatables-oracle": "^9.10",
-        "bugsnag/bugsnag-laravel": "^2.19"
+        "laravel/passport": "^10.0",
+        "laravel/ui": "^3.0",
+        "doctrine/dbal": "^2.11",
+        "laraveldaily/laravel-invoices": "^2.0",
+        "yajra/laravel-datatables-oracle": "^9.11",
+        "bugsnag/bugsnag-laravel": "^2.20"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.8",
         "fzaninotto/faker": "^1.9.1",
-        "mockery/mockery": "^1.4.1",
-        "nunomaduro/collision": "^4.2",
-        "phpunit/phpunit": "^9.3",
-        "laravel/dusk": "^6.4"
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.4",
+        "laravel/dusk": "^6.8"
     },
     "config": {
         "optimize-autoloader": true,
@@ -41,12 +41,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
