@@ -6,9 +6,9 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\MassDestroyInvoiceRequest;
 use App\Http\Requests\StoreInvoiceRequest;
 use App\Http\Requests\UpdateInvoiceRequest;
-use App\Invoice;
+use App\Models\Invoice;
 use App\Notifications\MonthlyMemberNotification;
-use App\Student;
+use App\Models\Student;
 use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Notification;
