@@ -5,7 +5,7 @@ return [
     'time_format'         => 'H:i:s',
     'primary_language'    => 'zh_TW',
     'available_languages' => [
-        'en' => 'English',
+        'en'    => 'English',
         'zh_TW' => 'Traditional Chinese'
     ],
 ];
