@@ -1,6 +1,8 @@
 <?php
 
-use App\Permission;
+namespace Database\Seeders;
+
+use App\Models\Permission;
 use Illuminate\Database\Seeder;
 
 class PermissionsTableSeeder extends Seeder
