@@ -1,6 +1,8 @@
 <?php
 
-use App\Role;
+namespace Database\Seeders;
+
+use App\Models\Role;
 use Illuminate\Database\Seeder;
 
 class RolesTableSeeder extends Seeder
